+++ xchainpy/xchainpy_util/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2021-08-12
+
+### Added 
+
+- Added Asset constants in asset.py (AssetBTC, AssetBCH and ...)
+
 ## 2021-08-5
 
 ### Added